// Arashyn-FE/src/lib/api/generated/models/grammar-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Arashyn API
 * REST API for Arashyn Platform
 *
 * The version of the OpenAPI document: v1.0
 * Contact: touman876@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Group } from './group';
// May contain unused imports in some cases
// @ts-ignore
import type { NoteCreateRequest } from './note-create-request';

export interface GrammarCreateRequest {
    'title': string;
    'language': GrammarCreateRequestLanguageEnum;
    'isPublic': boolean;
    'groups': Array<Group>;
    'notes'?: Array<NoteCreateRequest>;
    'filterIds'?: Array<string>;
}

export enum GrammarCreateRequestLanguageEnum {
    Vi = 'VI',
    En = 'EN',
    Ja = 'JA',
    Ko = 'KO',
    Zh = 'ZH'
}