// Arashyn-FE/src/lib/api/generated/models/group.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Arashyn API
 * REST API for Arashyn Platform
 *
 * The version of the OpenAPI document: v1.0
 * Contact: touman876@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ComponentCreateRequest } from './component-create-request';
// May contain unused imports in some cases
// @ts-ignore
import type { MeaningCreateBase } from './meaning-create-base';

export interface Group {
    'groupKey': number;
    'components': Array<ComponentCreateRequest>;
    'meanings'?: Array<MeaningCreateBase>;
}