// Arashyn-FE/src/lib/api/generated/models/meaning-create-base.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Arashyn API
 * REST API for Arashyn Platform
 *
 * The version of the OpenAPI document: v1.0
 * Contact: touman876@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ExampleCreateRequest } from './example-create-request';

export interface MeaningCreateBase {
    'content': string;
    'isPublic'?: boolean;
    'examples'?: Array<ExampleCreateRequest>;
}